ManagementNode/api/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
mariadb>=1.0.0
paho-mqtt>=1.6.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
